update code according to style guide
